flush telemtry
