+++ packages/next/src/server/dev/hot-reloader-webpack.ts
@@ -109,8 +109,7 @@ import {
 } from './hot-reloader-shared-utils'
 import { getMcpMiddleware } from '../mcp/get-mcp-middleware'
 import { setStackFrameResolver } from '../mcp/tools/utils/format-errors'
-import { getMcpTelemetryUsage } from '../mcp/get-or-create-mcp-server'
-import { eventMcpToolUsage } from '../../telemetry/events/build'
+import { recordMcpTelemetry } from '../mcp/mcp-telemetry-tracker'
 import { getFileLogger } from './browser-logs/file-logger'
 import type { ServerCacheStatus } from '../../next-devtools/dev-overlay/cache-indicator'
 
@@ -1813,15 +1812,7 @@ export default class HotReloaderWebpack implements NextJsHotReloaderInterface {
 
   public close() {
     // Report MCP telemetry if MCP server is enabled
-    if (this.config.experimental.mcpServer) {
-      const mcpUsages = getMcpTelemetryUsage()
-      if (mcpUsages.length > 0) {
-        const events = eventMcpToolUsage(mcpUsages)
-        for (const event of events) {
-          this.telemetry.record(event)
-        }
-      }
-    }
+    recordMcpTelemetry(this.telemetry)
 
     this.webpackHotMiddleware?.close()
   }
